+++ src/control/xml/XmlImageNode.h
@@ -21,7 +21,7 @@ public:
 public:
     void setImage(cairo_surface_t* img);
 
-    static cairo_status_t pngWriteFunction(XmlImageNode* image, const unsigned char* data, uint64_t length);
+    static cairo_status_t pngWriteFunction(XmlImageNode* image, const unsigned char* data, unsigned int length);
 
     virtual void writeOut(OutputStream* out);
