+++ src/util/serializing/ObjectOutputStream.cpp
@@ -58,7 +58,7 @@ void ObjectOutputStream::writeData(const void* data, int64_t len, int64_t width)
     }
 }
 
-static auto cairoWriteFunction(GString* string, const unsigned char* data, uint64_t length) -> cairo_status_t {
+static auto cairoWriteFunction(GString* string, const unsigned char* data, unsigned int length) -> cairo_status_t {
     g_string_append_len(string, reinterpret_cast<const gchar*>(data), length);
     return CAIRO_STATUS_SUCCESS;
 }
